+++ py3270/tests/test_emulator.py
@@ -1,4 +1,4 @@
-from cStringIO import StringIO
+from io import BytesIO
 import subprocess
 
 from blazeutils.testing import raises
@@ -12,11 +12,11 @@ from py3270 import EmulatorBase, Emulator, Command, Status, FieldTruncateError, 
 
 class MEmulator(Emulator):
     def __init__(self, visible=False, timeout=3):
-        Emulator.__init__(self, visible, timeout, _sp=mock.Mock())
+        Emulator.__init__(self, visible, timeout)
 
 
 class MCommand(Command):
-    def __init__(self, data='', result='ok', cmdstr='testcmd'):
+    def __init__(self, data='', result='ok', cmdstr=b'testcmd'):
         Command.__init__(self, ExamApp(data, result), cmdstr)
 
 
@@ -24,15 +24,18 @@ class ExamApp(object):
     """ An object with an interface like ExecutableApp used for testing """
 
     def __init__(self, data='', result='ok'):
-        self.stdin = StringIO()
+        self.stdin = BytesIO()
 
         response = data
         response += '' * 12 + '\n'
         response += result + '\n'
-        self.stdout = StringIO(response)
+        self.stdout = BytesIO(response.encode('ascii'))
 
     def connect(self, host):
         return False
+
+    def close(self):
+        self.stdin.close()
 
     def write(self, data):
         self.stdin.write(data)
@@ -50,15 +53,15 @@ class TestStatus(object):
         eq_(s.keyboard, None)
 
     def test_split(self):
-        statstr = 'U F U C(192.85.72.197) I 4 24 80 16 22 0x0 0.082'
+        statstr = b'U F U C(192.85.72.197) I 4 24 80 16 22 0x0 0.082'
         s = Status(statstr)
-        eq_(s.as_string, statstr)
-        eq_(s.keyboard, 'U')
+        eq_(s.as_string, statstr.decode('ascii'))
+        eq_(s.keyboard, b'U')
 
     def test_str_magic(self):
-        statstr = 'U F U C(192.85.72.197) I 4 24 80 16 22 0x0 0.082'
+        statstr = b'U F U C(192.85.72.197) I 4 24 80 16 22 0x0 0.082'
         s = Status(statstr)
-        eq_(str(s), 'STATUS: ' + statstr)
+        eq_(str(s), 'STATUS: ' + statstr.decode('ascii'))
 
 
 class TestEmulator(object):
@@ -69,13 +72,13 @@ class TestEmulator(object):
         assert em.last_host is None
         em.connect('localhost')
         assert em.last_host == 'localhost'
-        m_ec.assert_called_once_with('Connect(localhost)')
+        m_ec.assert_called_once_with(b'Connect(localhost)')
 
     @mock.patch('py3270.Emulator.exec_command')
     def test_terminate(self, m_ec):
         em = MEmulator()
         em.terminate()
-        m_ec.assert_called_once_with('Quit')
+        m_ec.assert_called_once_with(b'Quit')
         assert em.is_terminated
 
     @raises(TerminatedError)
@@ -90,31 +93,31 @@ class TestEmulator(object):
         em = MEmulator()
 
         em.send_enter()
-        m_ec.assert_called_with('Enter')
+        m_ec.assert_called_with(b'Enter')
 
         em.send_string('foo"bar')
-        m_ec.assert_called_with('String("foo\"bar")')
+        m_ec.assert_called_with(b'String("foo\"bar")')
 
         em.move_to(5, 7)
-        m_ec.assert_called_with('MoveCursor(4, 6)')
+        m_ec.assert_called_with(b'MoveCursor(4, 6)')
 
         em.send_pf3()
-        m_ec.assert_called_with('PF(3)')
+        m_ec.assert_called_with(b'PF(3)')
 
         em.send_pf4()
-        m_ec.assert_called_with('PF(4)')
+        m_ec.assert_called_with(b'PF(4)')
 
         em.send_pf5()
-        m_ec.assert_called_with('PF(5)')
+        m_ec.assert_called_with(b'PF(5)')
 
         em.send_pf6()
-        m_ec.assert_called_with('PF(6)')
+        m_ec.assert_called_with(b'PF(6)')
 
         m_ec.reset_mock()
         em.send_string('foobar', 5, 7)
         eq_(m_ec.call_args_list, [
-            mock.call('MoveCursor(4, 6)'),
-            mock.call('String("foobar")'),
+            mock.call(b'MoveCursor(4, 6)'),
+            mock.call(b'String("foobar")'),
         ])
 
     @mock.patch('py3270.Emulator.connect')
@@ -123,7 +126,7 @@ class TestEmulator(object):
         em = MEmulator()
         em.last_host = 'foo'
         em.reconnect()
-        m_ec.assert_called_once_with('Disconnect')
+        eq_(m_ec.call_args_list[0], mock.call(b'Disconnect'))
         eq_(m_connect.call_args_list, [
             # call from reconnect()
             mock.call('foo'),
@@ -135,32 +138,33 @@ class TestEmulator(object):
 
         em.fill_field(7, 9, 'foobar', 6)
         eq_(m_ec.call_args_list, [
-            mock.call('MoveCursor(6, 8)'),
-            mock.call('DeleteField'),
-            mock.call('String("foobar")'),
+            mock.call(b'MoveCursor(6, 8)'),
+            mock.call(b'DeleteField'),
+            mock.call(b'String("foobar")'),
         ])
         m_ec.reset_mock()
 
         em.fill_field(None, None, 'foobar', 6)
         eq_(m_ec.call_args_list, [
-            mock.call('DeleteField'),
-            mock.call('String("foobar")'),
+            mock.call(b'DeleteField'),
+            mock.call(b'String("foobar")'),
         ])
         m_ec.reset_mock()
 
         em.fill_field(1, None, 'foobar', 6)
         eq_(m_ec.call_args_list, [
-            mock.call('DeleteField'),
-            mock.call('String("foobar")'),
+            mock.call(b'DeleteField'),
+            mock.call(b'String("foobar")'),
         ])
 
     @mock.patch('py3270.Emulator.exec_command')
     def test_string_get(self, m_ec):
         em = MEmulator()
-        m_ec.return_value.data = ['foobar']
+        m_ec.return_value.data = [b'foobar']
         result = em.string_get(7,9,5)
         eq_(result, 'foobar')
-        m_ec.assert_called_with('Ascii(6,8,5)')
+        m_ec.assert_called_with(b'Ascii(6,8,5)')
+        em.terminate()
 
     @raises(AssertionError)
     @mock.patch('py3270.Emulator.exec_command')
@@ -187,43 +191,43 @@ class TestEmulator(object):
     @mock.patch('py3270.Emulator.exec_command')
     def test_wait_for_field(self, m_ec):
         em = MEmulator()
-        em.status.keyboard = 'U'
+        em.status.keyboard = b'U'
         em.wait_for_field()
 
-        m_ec.assert_called_once_with('Wait(3, InputField)')
+        m_ec.assert_called_once_with(b'Wait(3, InputField)')
 
     @mock.patch('py3270.Emulator.exec_command')
     def test_wait_for_field_custom_timeout(self, m_ec):
         em =  MEmulator(timeout=5)
-        em.status.keyboard = 'U'
+        em.status.keyboard = b'U'
         em.wait_for_field()
 
-        m_ec.assert_called_once_with('Wait(5, InputField)')
+        m_ec.assert_called_once_with(b'Wait(5, InputField)')
 
     @raises(KeyboardStateError, 'keyboard not unlocked, state was: E')
     @mock.patch('py3270.Emulator.exec_command')
     def test_wait_for_field_exception(self, m_ec):
         em = MEmulator()
-        em.status.keyboard = 'E'
+        em.status.keyboard = b'E'
         em.wait_for_field()
 
     @mock.patch('py3270.Emulator.exec_command')
     def test_not_is_connected(self, m_ec):
         em = MEmulator()
         def assign_status(*args, **kwargs):
-            em.status.connection_state = 'N'
+            em.status.connection_state = b'N'
         m_ec.side_effect = assign_status
         assert not em.is_connected()
-        m_ec.assert_called_once_with('ignore')
+        m_ec.assert_called_once_with(b'ignore')
 
     @mock.patch('py3270.Emulator.exec_command')
     def test_is_connected(self, m_ec):
         em = MEmulator()
         def assign_status(*args, **kwargs):
-            em.status.connection_state = 'C(192.168.1.1)'
+            em.status.connection_state = b'C(192.168.1.1)'
         m_ec.side_effect = assign_status
         assert em.is_connected()
-        m_ec.assert_called_once_with('ignore')
+        m_ec.assert_called_once_with(b'ignore')
 
     @raises(Exception, 'EmulatorBase has been replaced by Emulator.')
     def test_emulatorbase_exception(self):
@@ -234,7 +238,7 @@ class TestCommand(object):
 
     def test_defaults(self):
         cmd = MCommand()
-        eq_(cmd.cmdstr, 'testcmd')
+        eq_(cmd.cmdstr, b'testcmd')
         eq_(cmd.status_line, None)
         eq_(cmd.data, [])
 
@@ -242,21 +246,21 @@ class TestCommand(object):
     def test_stdin(self, m_write):
         cmd = MCommand()
         cmd.execute()
-        cmd.app.write.assert_called_once_with('testcmd\n')
+        cmd.app.write.assert_called_once_with(b'testcmd\n')
 
     def test_data(self):
         data = 'data: foo\n'
         data += 'data: bar\n'
         cmd = MCommand(data)
         cmd.execute()
-        eq_(cmd.data, ['foo', 'bar'])
+        eq_(cmd.data, [b'foo', b'bar'])
 
     def test_data_windows_line_endings(self):
         data = 'data: foo\r\n'
         data += 'data: bar\r\n'
         cmd = MCommand(data)
         cmd.execute()
-        eq_(cmd.data, ['foo', 'bar'])
+        eq_(cmd.data, [b'foo', b'bar'])
 
     @raises(CommandError, 'some kind of error')
     def test_error_response(self):
@@ -281,7 +285,7 @@ class TestCommand(object):
         cmd.execute()
 
     def test_blank_result_with_quit(self):
-        cmd = MCommand(result='', cmdstr='Quit')
+        cmd = MCommand(result='', cmdstr=b'Quit')
         # running without exception is sufficient for this test
         cmd.execute()